Feature/item

+++ weddingyou_backend/src/main/java/com/mysite/weddingyou_backend/userLogin/UserLogin.java
@@ -1,11 +1,16 @@
 package com.mysite.weddingyou_backend.userLogin;
 
+import java.io.Serializable;
 import java.time.LocalDateTime;
+
+import com.fasterxml.jackson.annotation.JsonManagedReference;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
@@ -18,7 +23,7 @@ public class UserLogin {
 	@Id // pk 지정
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "user_id")
-	private int userId;
+	private Long userId;
 
 	@Column(name = "name", nullable = false)
 	private String name;
